Improved the final state analysis; removed all output from the dumping code.
Currently the dumping code will call functions in the analysis code and expect
the analysis code to maintain it's own list containing the data. This is not an
ideal solution; eventually I'll modify the dump code to construct a linked list
which will be passed to the analysis code. The analysis code will then iterate
over the list and perform analysis on it.